--- serviceProviderNodeJSPassport/too/view.ejs
@@ -1,3 +1,0 @@
-  <div>
-    <%= FCUrl %>
-  </div>
